Fixed iterator bug
Bug occurs when system allows user to reenter their correct login details but is still invalid. This is because the iterator.hasNext() has no more to iterate through so it is null that's why it is returning false. The fix is to create a local iterator rather than a global iterator in the VerifyUser() function.